Move running VMs directory to proper level
