+++ src/views/clusteroverview/overview/components/running-vms-per-template-card/RunningVMsPerTemplateCard.scss
