+++ src/views/clusteroverview/overview/components/running-vms-per-template-card/utils/EmptyStateNoVMs.tsx
